rename: FpWeights 관련 코드 디렉토리 이동
