+++ src/main/kotlin/org/scaleadvisor/backend/project/application/port/repository/fpweights/CreateFpWeightsPort.kt
@@ -1,4 +1,4 @@
-package org.scaleadvisor.backend.project.application.port.repository
+package org.scaleadvisor.backend.project.application.port.repository.fpweights
 
 import org.scaleadvisor.backend.project.domain.FpWeights
